Hive: Refactor commit lock mechanism from HiveTableOperations (#5877)
