+++ hive-metastore/src/main/java/org/apache/iceberg/hive/HiveTableOperations.java
@@ -22,20 +22,12 @@ import static org.apache.iceberg.TableProperties.GC_ENABLED;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.github.benmanes.caffeine.cache.Cache;
-import com.github.benmanes.caffeine.cache.Caffeine;
-import java.net.InetAddress;
-import java.net.UnknownHostException;
 import java.util.Collections;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
-import java.util.UUID;
-import java.util.concurrent.Executors;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.ScheduledFuture;
-import java.util.concurrent.TimeUnit;
 import java.util.concurrent.locks.ReentrantLock;
 import java.util.stream.Collectors;
 import org.apache.hadoop.conf.Configuration;
@@ -43,17 +35,8 @@ import org.apache.hadoop.hive.common.StatsSetupConst;
 import org.apache.hadoop.hive.metastore.IMetaStoreClient;
 import org.apache.hadoop.hive.metastore.TableType;
 import org.apache.hadoop.hive.metastore.api.InvalidObjectException;
-import org.apache.hadoop.hive.metastore.api.LockComponent;
-import org.apache.hadoop.hive.metastore.api.LockLevel;
-import org.apache.hadoop.hive.metastore.api.LockRequest;
-import org.apache.hadoop.hive.metastore.api.LockResponse;
-import org.apache.hadoop.hive.metastore.api.LockState;
-import org.apache.hadoop.hive.metastore.api.LockType;
 import org.apache.hadoop.hive.metastore.api.NoSuchObjectException;
 import org.apache.hadoop.hive.metastore.api.SerDeInfo;
-import org.apache.hadoop.hive.metastore.api.ShowLocksRequest;
-import org.apache.hadoop.hive.metastore.api.ShowLocksResponse;
-import org.apache.hadoop.hive.metastore.api.ShowLocksResponseElement;
 import org.apache.hadoop.hive.metastore.api.StorageDescriptor;
 import org.apache.hadoop.hive.metastore.api.Table;
 import org.apache.hadoop.hive.metastore.api.hive_metastoreConstants;
@@ -75,16 +58,12 @@ import org.apache.iceberg.exceptions.ValidationException;
 import org.apache.iceberg.hadoop.ConfigProperties;
 import org.apache.iceberg.io.FileIO;
 import org.apache.iceberg.relocated.com.google.common.annotations.VisibleForTesting;
-import org.apache.iceberg.relocated.com.google.common.base.MoreObjects;
 import org.apache.iceberg.relocated.com.google.common.base.Preconditions;
 import org.apache.iceberg.relocated.com.google.common.collect.BiMap;
 import org.apache.iceberg.relocated.com.google.common.collect.ImmutableBiMap;
 import org.apache.iceberg.relocated.com.google.common.collect.ImmutableMap;
-import org.apache.iceberg.relocated.com.google.common.collect.Lists;
 import org.apache.iceberg.relocated.com.google.common.collect.Maps;
-import org.apache.iceberg.relocated.com.google.common.util.concurrent.ThreadFactoryBuilder;
 import org.apache.iceberg.util.JsonUtil;
-import org.apache.iceberg.util.Tasks;
 import org.apache.thrift.TException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -96,37 +75,15 @@ import org.slf4j.LoggerFactory;
 public class HiveTableOperations extends BaseMetastoreTableOperations {
   private static final Logger LOG = LoggerFactory.getLogger(HiveTableOperations.class);
 
-  private static final String HIVE_ACQUIRE_LOCK_TIMEOUT_MS = "iceberg.hive.lock-timeout-ms";
-  private static final String HIVE_LOCK_CHECK_MIN_WAIT_MS = "iceberg.hive.lock-check-min-wait-ms";
-  private static final String HIVE_LOCK_CHECK_MAX_WAIT_MS = "iceberg.hive.lock-check-max-wait-ms";
-  private static final String HIVE_LOCK_CREATION_TIMEOUT_MS =
-      "iceberg.hive.lock-creation-timeout-ms";
-  private static final String HIVE_LOCK_CREATION_MIN_WAIT_MS =
-      "iceberg.hive.lock-creation-min-wait-ms";
-  private static final String HIVE_LOCK_CREATION_MAX_WAIT_MS =
-      "iceberg.hive.lock-creation-max-wait-ms";
-  private static final String HIVE_LOCK_HEARTBEAT_INTERVAL_MS =
-      "iceberg.hive.lock-heartbeat-interval-ms";
   private static final String HIVE_ICEBERG_METADATA_REFRESH_MAX_RETRIES =
       "iceberg.hive.metadata-refresh-max-retries";
-  private static final String HIVE_TABLE_LEVEL_LOCK_EVICT_MS =
-      "iceberg.hive.table-level-lock-evict-ms";
-
   // the max size is based on HMS backend database. For Hive versions below 2.3, the max table
   // parameter size is 4000
   // characters, see https://issues.apache.org/jira/browse/HIVE-12274
   // set to 0 to not expose Iceberg metadata in HMS Table properties.
   private static final String HIVE_TABLE_PROPERTY_MAX_SIZE = "iceberg.hive.table-property-max-size";
   private static final long HIVE_TABLE_PROPERTY_MAX_SIZE_DEFAULT = 32672;
-  private static final long HIVE_ACQUIRE_LOCK_TIMEOUT_MS_DEFAULT = 3 * 60 * 1000; // 3 minutes
-  private static final long HIVE_LOCK_CHECK_MIN_WAIT_MS_DEFAULT = 50; // 50 milliseconds
-  private static final long HIVE_LOCK_CHECK_MAX_WAIT_MS_DEFAULT = 5 * 1000; // 5 seconds
-  private static final long HIVE_LOCK_CREATION_TIMEOUT_MS_DEFAULT = 3 * 60 * 1000; // 3 minutes
-  private static final long HIVE_LOCK_CREATION_MIN_WAIT_MS_DEFAULT = 50; // 50 milliseconds
-  private static final long HIVE_LOCK_CREATION_MAX_WAIT_MS_DEFAULT = 5 * 1000; // 5 seconds
-  private static final long HIVE_LOCK_HEARTBEAT_INTERVAL_MS_DEFAULT = 4 * 60 * 1000; // 4 minutes
   private static final int HIVE_ICEBERG_METADATA_REFRESH_MAX_RETRIES_DEFAULT = 2;
-  private static final long HIVE_TABLE_LEVEL_LOCK_EVICT_MS_DEFAULT = TimeUnit.MINUTES.toMillis(10);
   private static final BiMap<String, String> ICEBERG_TO_HMS_TRANSLATION =
       ImmutableBiMap.of(
           // gc.enabled in Iceberg and external.table.purge in Hive are meant to do the same things
@@ -134,13 +91,6 @@ public class HiveTableOperations extends BaseMetastoreTableOperations {
           GC_ENABLED, "external.table.purge");
 
   private static Cache<String, ReentrantLock> commitLockCache;
-
-  private static synchronized void initTableLevelLockCache(long evictionTimeout) {
-    if (commitLockCache == null) {
-      commitLockCache =
-          Caffeine.newBuilder().expireAfterAccess(evictionTimeout, TimeUnit.MILLISECONDS).build();
-    }
-  }
 
   /**
    * Provides key translation where necessary between Iceberg and HMS props. This translation is
@@ -161,28 +111,15 @@ public class HiveTableOperations extends BaseMetastoreTableOperations {
     return ICEBERG_TO_HMS_TRANSLATION.inverse().getOrDefault(hmsProp, hmsProp);
   }
 
-  private static class WaitingForLockException extends RuntimeException {
-    WaitingForLockException(String message) {
-      super(message);
-    }
-  }
-
   private final String fullName;
+  private final String catalogName;
   private final String database;
   private final String tableName;
   private final Configuration conf;
-  private final long lockAcquireTimeout;
-  private final long lockCheckMinWaitTime;
-  private final long lockCheckMaxWaitTime;
-  private final long lockCreationTimeout;
-  private final long lockCreationMinWaitTime;
-  private final long lockCreationMaxWaitTime;
-  private final long lockHeartbeatIntervalTime;
   private final long maxHiveTablePropertySize;
   private final int metadataRefreshMaxRetries;
   private final FileIO fileIO;
   private final ClientPool<IMetaStoreClient, TException> metaClients;
-  private final ScheduledExecutorService exitingScheduledExecutorService;
 
   protected HiveTableOperations(
       Configuration conf,
@@ -195,37 +132,15 @@ public class HiveTableOperations extends BaseMetastoreTableOperations {
     this.metaClients = metaClients;
     this.fileIO = fileIO;
     this.fullName = catalogName + "." + database + "." + table;
+    this.catalogName = catalogName;
     this.database = database;
     this.tableName = table;
-    this.lockAcquireTimeout =
-        conf.getLong(HIVE_ACQUIRE_LOCK_TIMEOUT_MS, HIVE_ACQUIRE_LOCK_TIMEOUT_MS_DEFAULT);
-    this.lockCheckMinWaitTime =
-        conf.getLong(HIVE_LOCK_CHECK_MIN_WAIT_MS, HIVE_LOCK_CHECK_MIN_WAIT_MS_DEFAULT);
-    this.lockCheckMaxWaitTime =
-        conf.getLong(HIVE_LOCK_CHECK_MAX_WAIT_MS, HIVE_LOCK_CHECK_MAX_WAIT_MS_DEFAULT);
-    this.lockCreationTimeout =
-        conf.getLong(HIVE_LOCK_CREATION_TIMEOUT_MS, HIVE_LOCK_CREATION_TIMEOUT_MS_DEFAULT);
-    this.lockCreationMinWaitTime =
-        conf.getLong(HIVE_LOCK_CREATION_MIN_WAIT_MS, HIVE_LOCK_CREATION_MIN_WAIT_MS_DEFAULT);
-    this.lockCreationMaxWaitTime =
-        conf.getLong(HIVE_LOCK_CREATION_MAX_WAIT_MS, HIVE_LOCK_CREATION_MAX_WAIT_MS_DEFAULT);
-    this.lockHeartbeatIntervalTime =
-        conf.getLong(HIVE_LOCK_HEARTBEAT_INTERVAL_MS, HIVE_LOCK_HEARTBEAT_INTERVAL_MS_DEFAULT);
     this.metadataRefreshMaxRetries =
         conf.getInt(
             HIVE_ICEBERG_METADATA_REFRESH_MAX_RETRIES,
             HIVE_ICEBERG_METADATA_REFRESH_MAX_RETRIES_DEFAULT);
     this.maxHiveTablePropertySize =
         conf.getLong(HIVE_TABLE_PROPERTY_MAX_SIZE, HIVE_TABLE_PROPERTY_MAX_SIZE_DEFAULT);
-    long tableLevelLockCacheEvictionTimeout =
-        conf.getLong(HIVE_TABLE_LEVEL_LOCK_EVICT_MS, HIVE_TABLE_LEVEL_LOCK_EVICT_MS_DEFAULT);
-    this.exitingScheduledExecutorService =
-        Executors.newSingleThreadScheduledExecutor(
-            new ThreadFactoryBuilder()
-                .setDaemon(true)
-                .setNameFormat("iceberg-hive-lock-heartbeat-%d")
-                .build());
-    initTableLevelLockCache(tableLevelLockCacheEvictionTimeout);
   }
 
   @Override
@@ -277,19 +192,10 @@ public class HiveTableOperations extends BaseMetastoreTableOperations {
 
     CommitStatus commitStatus = CommitStatus.FAILURE;
     boolean updateHiveTable = false;
-    String agentInfo = "Iceberg-" + UUID.randomUUID();
-    Optional<Long> lockId = Optional.empty();
-    // getting a process-level lock per table to avoid concurrent commit attempts to the same table
-    // from the same
-    // JVM process, which would result in unnecessary and costly HMS lock acquisition requests
-    ReentrantLock tableLevelMutex = commitLockCache.get(fullName, t -> new ReentrantLock(true));
-    tableLevelMutex.lock();
-    HiveLockHeartbeat hiveLockHeartbeat = null;
+
+    HiveLock lock = lockObject();
     try {
-      lockId = Optional.of(acquireLock(agentInfo));
-      hiveLockHeartbeat =
-          new HiveLockHeartbeat(metaClients, lockId.get(), lockHeartbeatIntervalTime);
-      hiveLockHeartbeat.schedule(exitingScheduledExecutorService);
+      lock.lock();
 
       Table tbl = loadHmsTable();
 
@@ -339,26 +245,20 @@ public class HiveTableOperations extends BaseMetastoreTableOperations {
         tbl.getParameters().remove(StatsSetupConst.COLUMN_STATS_ACCURATE);
       }
 
-      try {
-        if (hiveLockHeartbeat.future.isCancelled()
-            || hiveLockHeartbeat.encounteredException != null) {
-          throw new CommitFailedException(
-              "Failed to heartbeat for hive lock. %s",
-              hiveLockHeartbeat.encounteredException.getMessage());
-        }
+      lock.ensureActive();
 
+      try {
         persistTable(tbl, updateHiveTable);
-        if (hiveLockHeartbeat.future.isCancelled()
-            || hiveLockHeartbeat.encounteredException != null) {
-          throw new CommitStateUnknownException(
-              "Failed to heartbeat for hive lock while "
-                  + "committing changes. This can lead to a concurrent commit attempt be able to overwrite this commit. "
-                  + "Please check the commit history. If you are running into this issue, try reducing "
-                  + "iceberg.hive.lock-heartbeat-interval-ms.",
-              hiveLockHeartbeat.encounteredException);
-        }
+        lock.ensureActive();
 
         commitStatus = CommitStatus.SUCCESS;
+      } catch (LockException le) {
+        throw new CommitStateUnknownException(
+            "Failed to heartbeat for hive lock while "
+                + "committing changes. This can lead to a concurrent commit attempt be able to overwrite this commit. "
+                + "Please check the commit history. If you are running into this issue, try reducing "
+                + "iceberg.hive.lock-heartbeat-interval-ms.",
+            le);
       } catch (org.apache.hadoop.hive.metastore.api.AlreadyExistsException e) {
         throw new AlreadyExistsException(e, "Table already exists: %s.%s", database, tableName);
 
@@ -393,7 +293,7 @@ public class HiveTableOperations extends BaseMetastoreTableOperations {
             throw new CommitStateUnknownException(e);
         }
       }
-    } catch (TException | UnknownHostException e) {
+    } catch (TException e) {
       throw new RuntimeException(
           String.format("Metastore operation failed for %s.%s", database, tableName), e);
 
@@ -401,13 +301,11 @@ public class HiveTableOperations extends BaseMetastoreTableOperations {
       Thread.currentThread().interrupt();
       throw new RuntimeException("Interrupted during commit", e);
 
-    } finally {
-      if (hiveLockHeartbeat != null) {
-        hiveLockHeartbeat.cancel();
-      }
+    } catch (LockException e) {
+      throw new CommitFailedException(e);
 
-      cleanupMetadataAndUnlock(
-          commitStatus, newMetadataLocation, lockId, tableLevelMutex, agentInfo);
+    } finally {
+      cleanupMetadataAndUnlock(commitStatus, newMetadataLocation, lock);
     }
 
     LOG.info(
@@ -630,82 +528,8 @@ public class HiveTableOperations extends BaseMetastoreTableOperations {
     return storageDescriptor;
   }
 
-  @VisibleForTesting
-  long acquireLock(String agentInfo) throws UnknownHostException, TException, InterruptedException {
-    LockInfo lockInfo = tryLock(agentInfo);
-
-    final long start = System.currentTimeMillis();
-    long duration = 0;
-    boolean timeout = false;
-
-    try {
-      if (lockInfo.lockState.equals(LockState.WAITING)) {
-        // Retry count is the typical "upper bound of retries" for Tasks.run() function. In fact,
-        // the maximum number of
-        // attempts the Tasks.run() would try is `retries + 1`. Here, for checking locks, we use
-        // timeout as the
-        // upper bound of retries. So it is just reasonable to set a large retry count. However, if
-        // we set
-        // Integer.MAX_VALUE, the above logic of `retries + 1` would overflow into
-        // Integer.MIN_VALUE. Hence,
-        // the retry is set conservatively as `Integer.MAX_VALUE - 100` so it doesn't hit any
-        // boundary issues.
-        Tasks.foreach(lockInfo.lockId)
-            .retry(Integer.MAX_VALUE - 100)
-            .exponentialBackoff(lockCheckMinWaitTime, lockCheckMaxWaitTime, lockAcquireTimeout, 1.5)
-            .throwFailureWhenFinished()
-            .onlyRetryOn(WaitingForLockException.class)
-            .run(
-                id -> {
-                  try {
-                    LockResponse response = metaClients.run(client -> client.checkLock(id));
-                    LockState newState = response.getState();
-                    lockInfo.lockState = newState;
-                    if (newState.equals(LockState.WAITING)) {
-                      throw new WaitingForLockException(
-                          String.format("Waiting for lock on table %s.%s", database, tableName));
-                    }
-                  } catch (InterruptedException e) {
-                    Thread.interrupted(); // Clear the interrupt status flag
-                    LOG.warn(
-                        "Interrupted while waiting for lock on table {}.{}",
-                        database,
-                        tableName,
-                        e);
-                  }
-                },
-                TException.class);
-      }
-    } catch (WaitingForLockException waitingForLockException) {
-      timeout = true;
-      duration = System.currentTimeMillis() - start;
-    } finally {
-      if (!lockInfo.lockState.equals(LockState.ACQUIRED)) {
-        unlock(Optional.of(lockInfo.lockId), agentInfo);
-      }
-    }
-
-    // timeout and do not have lock acquired
-    if (timeout && !lockInfo.lockState.equals(LockState.ACQUIRED)) {
-      throw new CommitFailedException(
-          "Timed out after %s ms waiting for lock on %s.%s", duration, database, tableName);
-    }
-
-    if (!lockInfo.lockState.equals(LockState.ACQUIRED)) {
-      throw new CommitFailedException(
-          "Could not acquire the lock on %s.%s, lock request ended in state %s",
-          database, tableName, lockInfo.lockState);
-    }
-
-    return lockInfo.lockId;
-  }
-
   private void cleanupMetadataAndUnlock(
-      CommitStatus commitStatus,
-      String metadataLocation,
-      Optional<Long> lockId,
-      ReentrantLock tableLevelMutex,
-      String agentInfo) {
+      CommitStatus commitStatus, String metadataLocation, HiveLock lock) {
     try {
       if (commitStatus == CommitStatus.FAILURE) {
         // If we are sure the commit failed, clean up the uncommitted metadata file
@@ -714,62 +538,8 @@ public class HiveTableOperations extends BaseMetastoreTableOperations {
     } catch (RuntimeException e) {
       LOG.error("Failed to cleanup metadata file at {}", metadataLocation, e);
     } finally {
-      unlock(lockId, agentInfo);
-      tableLevelMutex.unlock();
+      lock.unlock();
     }
-  }
-
-  private void unlock(Optional<Long> lockId, String agentInfo) {
-    Long id = null;
-    try {
-      if (!lockId.isPresent()) {
-        // Try to find the lock based on agentInfo. Only works with Hive 2 or later.
-        if (HiveVersion.min(HiveVersion.HIVE_2)) {
-          LockInfo lockInfo = findLock(agentInfo);
-          if (lockInfo == null) {
-            // No lock found
-            LOG.info("No lock found with {} agentInfo", agentInfo);
-            return;
-          }
-
-          id = lockInfo.lockId;
-        } else {
-          LOG.warn("Could not find lock with HMSClient {}", HiveVersion.current());
-          return;
-        }
-      } else {
-        id = lockId.get();
-      }
-
-      doUnlock(id);
-    } catch (InterruptedException ie) {
-      if (id != null) {
-        // Interrupted unlock. We try to unlock one more time if we have a lockId
-        try {
-          Thread.interrupted(); // Clear the interrupt status flag for now, so we can retry unlock
-          LOG.warn("Interrupted unlock we try one more time {}.{}", database, tableName, ie);
-          doUnlock(id);
-        } catch (Exception e) {
-          LOG.warn("Failed to unlock even on 2nd attempt {}.{}", database, tableName, e);
-        } finally {
-          Thread.currentThread().interrupt(); // Set back the interrupt status
-        }
-      } else {
-        Thread.currentThread().interrupt(); // Set back the interrupt status
-        LOG.warn("Interrupted finding locks to unlock {}.{}", database, tableName, ie);
-      }
-    } catch (Exception e) {
-      LOG.warn("Failed to unlock {}.{}", database, tableName, e);
-    }
-  }
-
-  @VisibleForTesting
-  void doUnlock(long lockId) throws TException, InterruptedException {
-    metaClients.run(
-        client -> {
-          client.unlock(lockId);
-          return null;
-        });
   }
 
   static void validateTableIsIceberg(Table table, String fullName) {
@@ -808,169 +578,8 @@ public class HiveTableOperations extends BaseMetastoreTableOperations {
         ConfigProperties.ENGINE_HIVE_ENABLED, TableProperties.ENGINE_HIVE_ENABLED_DEFAULT);
   }
 
-  /**
-   * Tries to create a lock. If the lock creation fails, and it is possible then retries the lock
-   * creation a few times. If the lock creation is successful then a {@link LockInfo} is returned,
-   * otherwise an appropriate exception is thrown.
-   *
-   * @param agentInfo The agentInfo which should be used during lock creation
-   * @return The created lock
-   * @throws UnknownHostException When we are not able to fill the hostname for lock creation
-   * @throws TException When there is an error during lock creation
-   */
-  @SuppressWarnings("ReverseDnsLookup")
-  private LockInfo tryLock(String agentInfo) throws UnknownHostException, TException {
-    LockInfo lockInfo = new LockInfo();
-
-    final LockComponent lockComponent =
-        new LockComponent(LockType.EXCLUSIVE, LockLevel.TABLE, database);
-    lockComponent.setTablename(tableName);
-    final LockRequest lockRequest =
-        new LockRequest(
-            Lists.newArrayList(lockComponent),
-            System.getProperty("user.name"),
-            InetAddress.getLocalHost().getHostName());
-
-    // Only works in Hive 2 or later.
-    if (HiveVersion.min(HiveVersion.HIVE_2)) {
-      lockRequest.setAgentInfo(agentInfo);
-    }
-
-    Tasks.foreach(lockRequest)
-        .retry(Integer.MAX_VALUE - 100)
-        .exponentialBackoff(
-            lockCreationMinWaitTime, lockCreationMaxWaitTime, lockCreationTimeout, 2.0)
-        .shouldRetryTest(e -> e instanceof TException && HiveVersion.min(HiveVersion.HIVE_2))
-        .throwFailureWhenFinished()
-        .run(
-            request -> {
-              try {
-                LockResponse lockResponse = metaClients.run(client -> client.lock(request));
-                lockInfo.lockId = lockResponse.getLockid();
-                lockInfo.lockState = lockResponse.getState();
-              } catch (TException te) {
-                LOG.warn("Failed to acquire lock {}", request, te);
-                try {
-                  // If we can not check for lock, or we do not find it, then rethrow the exception
-                  // Otherwise we are happy as the findLock sets the lockId and the state correctly
-                  if (!HiveVersion.min(HiveVersion.HIVE_2)) {
-                    LockInfo lockFound = findLock(agentInfo);
-                    if (lockFound != null) {
-                      lockInfo.lockId = lockFound.lockId;
-                      lockInfo.lockState = lockFound.lockState;
-                      LOG.info("Found lock {} by agentInfo {}", lockInfo, agentInfo);
-                      return;
-                    }
-                  }
-
-                  throw te;
-                } catch (InterruptedException e) {
-                  Thread.currentThread().interrupt();
-                  LOG.warn(
-                      "Interrupted while checking for lock on table {}.{}", database, tableName, e);
-                  throw new RuntimeException("Interrupted while checking for lock", e);
-                }
-              } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-                LOG.warn("Interrupted while acquiring lock on table {}.{}", database, tableName, e);
-                throw new RuntimeException("Interrupted while acquiring lock", e);
-              }
-            },
-            TException.class);
-
-    // This should be initialized always, or exception should be thrown.
-    LOG.debug("Lock {} created for table {}.{}", lockInfo, database, tableName);
-    return lockInfo;
-  }
-
-  /**
-   * Search for the locks using HMSClient.showLocks identified by the agentInfo. If the lock is
-   * there, then a {@link LockInfo} object is returned. If the lock is not found <code>null</code>
-   * is returned.
-   *
-   * @param agentInfo The key for searching the locks
-   * @return The {@link LockInfo} for the found lock, or <code>null</code> if nothing found
-   */
-  private LockInfo findLock(String agentInfo) throws TException, InterruptedException {
-    Preconditions.checkArgument(
-        HiveVersion.min(HiveVersion.HIVE_2),
-        "Minimally Hive 2 HMS client is needed to find the Lock using the showLocks API call");
-    ShowLocksRequest showLocksRequest = new ShowLocksRequest();
-    showLocksRequest.setDbname(database);
-    showLocksRequest.setTablename(tableName);
-    ShowLocksResponse response = metaClients.run(client -> client.showLocks(showLocksRequest));
-    for (ShowLocksResponseElement lock : response.getLocks()) {
-      if (lock.getAgentInfo().equals(agentInfo)) {
-        // We found our lock
-        return new LockInfo(lock.getLockid(), lock.getState());
-      }
-    }
-
-    // Not found anything
-    return null;
-  }
-
-  private static class HiveLockHeartbeat implements Runnable {
-    private final ClientPool<IMetaStoreClient, TException> hmsClients;
-    private final long lockId;
-    private final long intervalMs;
-    private ScheduledFuture<?> future;
-    private volatile Exception encounteredException = null;
-
-    HiveLockHeartbeat(
-        ClientPool<IMetaStoreClient, TException> hmsClients, long lockId, long intervalMs) {
-      this.hmsClients = hmsClients;
-      this.lockId = lockId;
-      this.intervalMs = intervalMs;
-      this.future = null;
-    }
-
-    @Override
-    public void run() {
-      try {
-        hmsClients.run(
-            client -> {
-              client.heartbeat(0, lockId);
-              return null;
-            });
-      } catch (TException | InterruptedException e) {
-        this.encounteredException = e;
-        throw new CommitFailedException(e, "Failed to heartbeat for lock: %d", lockId);
-      }
-    }
-
-    public void schedule(ScheduledExecutorService scheduler) {
-      future =
-          scheduler.scheduleAtFixedRate(this, intervalMs / 2, intervalMs, TimeUnit.MILLISECONDS);
-    }
-
-    public void cancel() {
-      if (future != null) {
-        future.cancel(false);
-      }
-    }
-  }
-
-  private static class LockInfo {
-    private long lockId;
-    private LockState lockState;
-
-    private LockInfo() {
-      this.lockId = -1;
-      this.lockState = null;
-    }
-
-    private LockInfo(long lockId, LockState lockState) {
-      this.lockId = lockId;
-      this.lockState = lockState;
-    }
-
-    @Override
-    public String toString() {
-      return MoreObjects.toStringHelper(this)
-          .add("lockId", lockId)
-          .add("lockState", lockState)
-          .toString();
-    }
+  @VisibleForTesting
+  HiveLock lockObject() {
+    return new MetastoreLock(conf, metaClients, catalogName, database, tableName);
   }
 }
